Fix stats page error
Also cleaned up a few bits and pieces here and there while I was trying
to figure out what was going on. It turns that filtering a bunch of
FileResources based on their library type was creating duplicate
FileResources in the queryset (why?). Regardless, adding a distinct() at
the end of the queryset solved the issue.