Swap out django-rest-swagger for drf-yasg
For reference:

+ django-rest-swagger: https://github.com/marcgibbons/django-rest-swagger
+ drf-yasg: https://github.com/axnsan12/drf-yasg

Both do the same thing roughly but drf-yasg does it better (their
documentation explains this well), especially when it comes to
conforming properly to the OpenAPI spec.

The biggest benefits of drf-yasg to us are (1) that we get validators
(automatic, I think) to guarantee that we follow the OpenAPI 2 spec,
which is valuable for swagger-codegen
(https://github.com/swagger-api/swagger-codegen), if we end up using
that; and (2) that we get the very nice ReDoc interface for our API (the
one that Juan showed off to us).

Two things to note are that we need to upgrade Django from 1.11.4 to
1.11.13, which should be fine since those minor versions should only be
for bugfixes. We also need to upgrade Django REST Framework from 3.6.3
to 3.8.2. Again, I'm not too worried here, since our usage of DRF is
fairly pedestrian, so it's very unlikely anything would change as a
result of this on our end.